third_party/WebKit/LayoutTests/http/tests/serviceworker/fetch-body-mixin.html

<!DOCTYPE html>
<title>Service Worker: Body mixin</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
async_test(function(t) {
    var scope = 'resources/blank.html';
    service_worker_unregister_and_register(
        t, 'resources/fetch-body-mixin-worker.js', scope)
      .then(function(registration) {
          return wait_for_update(t, registration);
        })
      .then(function(sw) {
          var messageChannel = new MessageChannel();
          messageChannel.port1.onmessage = t.step_func(onMessage);
          sw.postMessage({port: messageChannel.port2}, [messageChannel.port2]);
        })
      .catch(unreached_rejection(t));

    var result = [];
    var expected = [
      'Text: <!DOCTYPE html>\n',
      'JSON: 1 2',
      'JSON: SyntaxError',
      'Blob: 16 : text/html',
      'ArrayBuffer: 16',
      '<!DOCTYPE html>\n : TypeError',
    ];

    function onMessage(e) {
      var message = e.data;
      if (message == 'quit') {
        assert_array_equals(result, expected,
                            'Worker should post back expected values.');
        service_worker_unregister_and_done(t, scope);
      } else {
        result.push(message);
      }
    }
  }, 'Verify FetchBodyStream in a Service Worker');
</script>
